Move ProgramEnrolment into the Patient directory
