+++ client/packages/system/src/Patient/ProgramEnrolment/DetailView/DetailView.tsx
