+++ client/packages/system/src/Patient/ProgramEnrolment/DetailView/index.ts
