+++ client/packages/system/src/Patient/ProgramEnrolment/api/index.ts
@@ -1,0 +1,2 @@
+export * from './hooks';
+export { ProgramFragment } from './operations.generated';
